requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
Pygments==2.6.1
pytz==2020.1
sqlparse==0.3.1
dj-database-url==0.5.0
django-jsonfield==1.4.0
gunicorn==20.0.4
six==1.15.0
